reset and resetPW almost done
admin can put new attractions into database but it can't put them on the website yet. For some reason the controller isn't comparing the new password in the database to the one the user is giving after reset. I can't get the index or view to recognize the variables I've set the attraction database with.